add pagination

+++ repository/contacts/removeContact.js
@@ -1,6 +1,6 @@
 const {Contact} = require('../../models/contacts')
-const removeContact = async (contactId) => {
-  const result = await Contact.findOneAndRemove({_id: contactId})
+const removeContact = async (contactId, user) => {
+  const result = await Contact.findOneAndRemove({_id: contactId, owner: user.id})
   return result
 }
